// examples/bookstore/generated/sql/mutations.ts
export type findOrderById_Row = {
  orderId: string;
  orderNumber: string;
  orderStatus: string;
  currency: string;
  placedAt: string;
  paidAt: string | null;
  shippedAt: string | null;
  shippingMethod: string | null;
  giftMessage: string | null;
};

export type findOrderById_Output = findOrderById_Row | null;
// ...
export type findOrderByNumber_Row = {
  orderId: string;
  orderNumber: string;
  orderStatus: string;
  currency: string;
  placedAt: string;
  paidAt: string | null;
  shippedAt: string | null;
  shippingMethod: string | null;
  giftMessage: string | null;
};

export type findOrderByNumber_Output = findOrderByNumber_Row | null;
// ...
export type createOrderItems_Input = {
  items: readonly [{ orderId: string; bookId: string; quantity: number; unitPrice: string; discountAmount: string | null }, ...{ orderId: string; bookId: string; quantity: number; unitPrice: string; discountAmount: string | null }[]];
};